Retrieve Data from BitBucket Asynchronously
This commit rewrites the way data is retrieved from BitBucket. Before
requests to BitBucket were made synchronously to always get the latest
state of the pull requests.

However, this was slow and the data were updated only every few minutes
anyway. Furthermore this meant that some additional data could not be
retrieved at all as this would have taken to long.

Now a worker will be spawned automatically which will retrieved the
latest data every minute and cache them for the UI to display. This
allows to make slower requests and get more data.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>